[MBL-15269][Student] Navigation drawer redesign
refs: MBL-15269
affects: Student
release note: Menu redesign

test plan: Generally check that all the menu items are working as intended.
