+++ apps/student/src/androidTest/java/com/instructure/student/ui/interaction/NavigationDrawerInteractionTest.kt
@@ -15,10 +15,18 @@
  */
 package com.instructure.student.ui.interaction
 
+import android.app.Activity
+import android.app.Instrumentation
+import android.content.Intent
 import android.os.Build
+import android.util.Log
+import androidx.test.espresso.intent.Intents
+import androidx.test.espresso.intent.matcher.IntentMatchers
+import androidx.test.espresso.web.webdriver.Locator
 import com.instructure.canvas.espresso.Stub
 import com.instructure.canvas.espresso.mockCanvas.MockCanvas
 import com.instructure.canvas.espresso.mockCanvas.init
+import com.instructure.canvasapi2.models.Course
 import com.instructure.canvasapi2.models.User
 import com.instructure.canvasapi2.utils.ApiPrefs
 import com.instructure.canvasapi2.utils.ContextKeeper
@@ -28,8 +36,11 @@ import com.instructure.panda_annotations.FeatureCategory
 import com.instructure.panda_annotations.Priority
 import com.instructure.panda_annotations.TestCategory
 import com.instructure.panda_annotations.TestMetaData
+import com.instructure.student.ui.pages.WebViewTextCheck
 import com.instructure.student.ui.utils.StudentTest
 import com.instructure.student.ui.utils.tokenLogin
+import org.hamcrest.CoreMatchers
+import org.junit.Before
 import org.junit.Test
 
 class NavigationDrawerInteractionTest : StudentTest() {
@@ -37,6 +48,18 @@ class NavigationDrawerInteractionTest : StudentTest() {
 
     private lateinit var student1: User
     private lateinit var student2: User
+    private lateinit var course: Course
+
+    private lateinit var activity: Activity
+
+    @Before
+    fun setUp() {
+        // If we try to read this later, it may be null, possibly because we will have navigated
+        // away from our initial activity.
+        activity = activityRule.activity
+
+
+    }
 
     // Should be able to change the user from the navigation drawer
     @Test
@@ -106,17 +129,153 @@ class NavigationDrawerInteractionTest : StudentTest() {
         val data = MockCanvas.init(
                 studentCount = 2,
                 courseCount = 1,
-                favoriteCourseCount = 1,
-                teacherCount = 1
+                favoriteCourseCount = 1
         )
 
         student1 = data.students.first()
         student2 = data.students.last()
+
+        course = data.courses.values.first()
 
         val token = data.tokenFor(student1)!!
         tokenLogin(data.domain, token, student1)
         dashboardPage.waitForRender()
 
         return data
+    }
+
+    // Should open a dialog and send a question for the selected course
+    // (Checks to see that we can fill out the question and the SEND button exists.)
+    @Test
+    @TestMetaData(Priority.P0, FeatureCategory.SETTINGS, TestCategory.INTERACTION, false)
+    fun testHelp_askQuestion() {
+
+        signInStudent()
+
+        dashboardPage.goToHelp()
+        helpPage.verifyAskAQuestion(course, "Here's a question")
+    }
+
+    // Should open the Canvas guides in a WebView
+    @Test
+    @TestMetaData(Priority.P0, FeatureCategory.SETTINGS, TestCategory.INTERACTION, false)
+    fun testHelp_searchCanvasGuides() {
+        signInStudent()
+
+        dashboardPage.goToHelp()
+        helpPage.launchGuides()
+        canvasWebViewPage.runTextChecks(
+                // Potentially brittle -- the web content could be changed by another team
+                WebViewTextCheck(Locator.CLASS_NAME, "lia-panel-heading-bar-title", "Guides by Product", 25)
+        )
+    }
+
+    // Should send an error report
+    // (Checks to see that we can fill out an error report and that the SEND button is displayed.)
+    @Test
+    @TestMetaData(Priority.P0, FeatureCategory.SETTINGS, TestCategory.INTERACTION, false)
+    fun testHelp_reportAProblem() {
+
+        signInStudent()
+
+        dashboardPage.goToHelp()
+        helpPage.verifyReportAProblem("Problem", "It's a problem!")
+    }
+
+    // Should send a pre-filled email intent. Should be addressed to mobilesupport@instructure.com.
+    //
+    // There is a LOT of aspirational code here, in that we would like to be able to handle
+    // an intent for a specific email app if one is present.  However, our app always launches
+    // an email app chooser, even if there is only one option.
+    //
+    // So this is a watered-down test that just checks whether an email app chooser gets displayed.
+    @Test
+    @TestMetaData(Priority.P0, FeatureCategory.SETTINGS, TestCategory.INTERACTION, false)
+    fun testHelp_submitFeatureIdea() {
+        signInStudent()
+
+        dashboardPage.goToHelp()
+
+        // Figure out which email apps we have installed on the device
+        var pkgMgr = activity.packageManager
+        var intent = Intent(Intent.ACTION_SEND)
+        intent.type = "message/rfc822"
+        val activities = pkgMgr.queryIntentActivities(intent, 0)
+        val matchedChooserActivities = activities.count()
+        for (activity in activities) {
+            Log.d("submitFeatureIdea","Resolved activity = $activity")
+        }
+
+        Intents.init()
+        try {
+            // Try to formulate what an email app chooser intent would look like, and how we might resolve it
+            val chooserIntentMatcher = IntentMatchers.hasAction(Intent.ACTION_CHOOSER)
+            val expectedChooserIntent = Intent(Intent.ACTION_SEND)
+            expectedChooserIntent.type = "message/rfc822"
+            expectedChooserIntent.`package` = "com.google.android.gm"
+
+            // Formulate what an actual email intent (NOT a chooser intent) would look like
+            val emailIntentMatcher = CoreMatchers.allOf(
+                    IntentMatchers.hasAction(Intent.ACTION_SEND),
+                    IntentMatchers.hasType("message/rfc822"),
+                    CoreMatchers.anyOf(
+                            IntentMatchers.hasExtra(Intent.EXTRA_EMAIL, arrayOf("support@instructure.com")),
+                            IntentMatchers.hasExtra(Intent.EXTRA_EMAIL, arrayOf("mobilesupport@instructure.com"))
+                    )
+            )
+
+            // Set up our intent catchers
+            Intents.intending(chooserIntentMatcher).respondWith(Instrumentation.ActivityResult(0, expectedChooserIntent))
+            Intents.intending(emailIntentMatcher).respondWith(Instrumentation.ActivityResult(0, null))
+
+            // Press the "Submit Feature" button
+            helpPage.submitFeature()
+
+//            // Depending on how many activities matched our email intent...
+//            if(matchedChooserActivities > 1) {
+//                // If multiple, just check that the chooser appeared.
+//                Intents.intended(chooserIntentMatcher)
+//            }
+//            else if(matchedChooserActivities == 1){
+//                // If single, check that our email intent was dispatched
+//                Intents.intended(emailIntentMatcher)
+//            }
+//            else {
+//                // If none, there is nothing much to do here
+//                Log.d("submitFeatureIdea","Not matched activities for SEND on device!")
+//            }
+
+            // :-( Our production code creates a chooser every time, even if there is only one email app option...
+            Intents.intended(chooserIntentMatcher)
+        }
+        finally {
+            Intents.release()
+        }
+    }
+
+    // Should send an intent to open the listing for Student App in the Play Store
+    @Test
+    @TestMetaData(Priority.P0, FeatureCategory.SETTINGS, TestCategory.INTERACTION, false)
+    fun testHelp_shareYourLove() {
+        signInStudent()
+
+        dashboardPage.goToHelp()
+        Intents.init()
+        try {
+            val expectedIntent = CoreMatchers.allOf(
+                    IntentMatchers.hasAction(Intent.ACTION_VIEW),
+                    CoreMatchers.anyOf(
+                            // Could be either of these, depending on whether the play store app is installed
+                            IntentMatchers.hasData("market://details?id=com.instructure.candroid"),
+                            IntentMatchers.hasData("https://play.google.com/store/apps/details?id=com.instructure.candroid")
+                    )
+            )
+            Intents.intending(expectedIntent).respondWith(Instrumentation.ActivityResult(0, null))
+            helpPage.shareYourLove()
+            Intents.intended(expectedIntent)
+        }
+        finally {
+            Intents.release()
+        }
     }
 }
