+++ apps/student/src/androidTest/java/com/instructure/student/ui/pages/DashboardPage.kt
@@ -164,6 +164,11 @@ class DashboardPage : BasePage(R.id.dashboardPage) {
         onViewWithId(R.id.navigationDrawerItem_changeUser).scrollTo().click()
     }
 
+    fun goToHelp() {
+        onView(hamburgerButtonMatcher).click()
+        onViewWithId(R.id.navigationDrawerItem_help).scrollTo().click()
+    }
+
     fun gotoGlobalFiles() {
         onView(hamburgerButtonMatcher).click()
         onViewWithId(R.id.navigationDrawerItem_files).click()
