+++ apps/student/src/main/res/values/styles.xml
@@ -153,8 +153,8 @@
         <item name="android:layout_width">match_parent</item>
         <item name="android:minHeight">?android:listPreferredItemHeightSmall</item>
         <item name="android:background">?android:selectableItemBackground</item>
-        <item name="android:paddingStart">16dp</item>
-        <item name="android:paddingEnd">16dp</item>
+        <item name="android:paddingStart">20dp</item>
+        <item name="android:paddingEnd">20dp</item>
         <item name="android:gravity">center_vertical</item>
         <item name="android:tint">@color/navigationDrawerTextDark</item>
     </style>
